requirements.txt
openai
pyautogui
playwright
browser-use
pillow
python-dotenv
